+++ native/src/seal/context.h
@@ -127,12 +127,12 @@ namespace seal
         /**
         Returns the name of parameter_error.
         */
-        SEAL_NODISCARD const std::string parameter_error_name() const;
+        SEAL_NODISCARD const std::string parameter_error_name() const noexcept;
 
         /**
         Returns a comprehensive message that interprets parameter_error.
         */
-        SEAL_NODISCARD const std::string parameter_error_message() const;
+        SEAL_NODISCARD const std::string parameter_error_message() const noexcept;
 
         /**
         Tells whether parameter_error is error_type::success.